Calculate area for all vertices
The reason gravity wasn't smooth was that initial area wasn't calculated
for non-base layers. Therefore, adding layers to the terrain caused
uninitialized (hence zero) areas, which blew up when trying to apply
gravity (which affects the terrain now as vertex weight divided by
initial area, i.e. something similar to thickness, an infinitesimally
thin column of terrain). That makes it comparable to buoyancy, which
acts on the base layer only with a dimensionless force (or, similarly,
which can be thought of as acting on the same infinitesimal column).

These adjustments should make the terrain generation somewhat
independent of the number of terrain divisions. However, it is still
dependent on the actual thickness of all layers together. That is kind
of desired, as mountains do need to be taller (in equilibrium) than
plains, so some degree of freedom is required. Given that we increase
buoyancy linearly with depth, equilibrium will always be attained at a
more or less similar altitude, at least so long as the buoyancy gradient
(the increase in force with decreasing altitude) compensates for the
increasing weight of the overlying terrain. Given that the Layer
currently assigns each vertex a weight corresponding to the volume that
it represents, a buoyancy gradient of 1.0 should accomplish that.